+++ lib/pages/file_manager_page.dart
@@ -14,6 +14,7 @@ import 'package:remote_storage/models/paged_listings.dart';
 import 'package:remote_storage/models/remote_storage_config.dart';
 import 'package:remote_storage/models/s3_objects.dart';
 import 'package:remote_storage/models/trash_item.dart';
+import 'package:remote_storage/platform/platform_info.dart';
 import 'package:remote_storage/services/file_access_service.dart';
 import 'package:remote_storage/services/file_access_transfer_request.dart';
 import 'package:remote_storage/services/desktop_file_transfer_service.dart';